Debouncing
For whatever reason the keyfobs habitually repeat transmissions. Perhaps they do it on purpose to ensure the message is received for alarm setting.

This is now debounced with a pretty long hysteresis of 8 seconds, because the last thing you want to do is use a button as a toggle, only turn your alarm system on and immediately off again.