Library methods and healthStatus implemented
I honestly don't know whether to keep this plug in the Tuya pile or not. The TS011F is a family of Tuya devices, all smart plugs or smart sockets and even DIN rail switches. But I only have access to the one sold through Lidl in the UK, so calling this a Tuya driver is a little disingenuous.