Fixed child creation at configureSpecifics()
This is a weird one. Pairing a new extension showed that the for loop to create the child devices was not being run, but instead was skipped without error. Switch from evaluating equivalence with '==' to '<=' and it now works.

Then also fixed the creation fault which didn't include the full driver title, which is now required for system drivers.