Added mqttGetStateType()
This is important as we don't initially know how many controllable switches a device has available, but Z2M changes the name of the state ("state" or "state_lX") when it's more than one.

We record this in the device network ID (DNI) as the penultimate integer, so we can know how many switches / states to expect ahead of time and use the right convention.